updated import statement in line with file split-up /name change
