+++ src/oogeso/core/devices/compressor.py
@@ -2,7 +2,7 @@ import pyomo.environ as pyo
 
 from oogeso.core.networks.network_node import NetworkNode
 from . import Device
-from oogeso.dto.oogeso_input_data_objects import (
+from oogeso.dto import (
     CarrierData,
     DeviceCompressor_elData,
     DeviceCompressor_gasData,
