+++ src/oogeso/core/devices/device.py
@@ -2,7 +2,7 @@ import pyomo.environ as pyo
 import logging
 from oogeso.core.networks.network import Network
 from oogeso.core.networks.network_node import NetworkNode
-from oogeso.dto.oogeso_input_data_objects import (
+from oogeso.dto import (
     CarrierData,
     DeviceData,
     OptimisationParametersData,
