+++ src/oogeso/dto/__init__.py
@@ -1,2 +1,3 @@
 from .oogeso_input_data_objects import *
+from .oogeso_input_data_objects_multienergy import *
 from .oogeso_output_data_objects import *
