+++ src/oogeso/dto/oogeso_input_data_objects.py
@@ -54,7 +54,7 @@ class DevicePowersourceData(DeviceData):
 
 
 @dataclass
-class DeviceSink_elData(DeviceData):
+class DevicePowersinkData(DeviceData):
     pass
